backend/app/requirements.txt
fastapi==0.111.1
uvicorn==0.30.3
psycopg2-binary
requests
